add analytics to most of the policy commands
