+++ pkg/policy/policy.go
@@ -4,8 +4,6 @@ import (
 	"crypto/sha256"
 	"fmt"
 	"strings"
-
-	"github.com/cloudquery/cloudquery/internal/getter"
 )
 
 type Policies []*Policy
@@ -44,19 +42,6 @@ type Check struct {
 	Reason       string    `hcl:"reason,optional"`
 }
 
-type Analytic struct {
-	// Whether policy will persist in database
-	Persistence bool
-	// Name of the policy
-	Name string
-	// Type of the policy i.e S3/Hub/Git
-	Type string
-	// The selector used for the policy
-	Selector string
-	// Whether policy is private
-	Private bool
-}
-
 type Policy struct {
 	// Name of the policy
 	Name string `hcl:"name,label"`
@@ -91,29 +76,6 @@ func (pp Policies) All() []string {
 		policyNames[i] = p.Name
 	}
 	return policyNames
-}
-
-func (p Policy) Analytic(dbPersistence bool) Analytic {
-	pa := Analytic{
-		Persistence: dbPersistence,
-		Name:        p.Name,
-		Type:        p.SourceType(),
-		Selector:    p.SubPolicy(),
-		Private:     p.SourceType() != "hub",
-	}
-	if !p.HasMeta() {
-		policyName, subPath := getter.ParseSourceSubPolicy(p.Source)
-		dp, _, _ := DetectPolicy(policyName, subPath)
-		pa.Type = dp.SourceType()
-		pa.Selector = subPath
-		pa.Name = policyName
-		pa.Private = p.SourceType() != "hub"
-	}
-	if pa.Private {
-		pa.Selector = "private"
-		pa.Name = p.Sha256Hash()
-	}
-	return pa
 }
 
 func (p Policy) String() string {
@@ -208,14 +170,4 @@ func (p Policy) Sha256Hash() string {
 	h := sha256.New()
 	h.Write([]byte(fmt.Sprintf("%v", p.Policies)))
 	return fmt.Sprintf("%x", h.Sum(nil))
-}
-
-func (a Analytic) Properties() map[string]interface{} {
-	return map[string]interface{}{
-		"policy_persistence": a.Persistence,
-		"policy_name":        a.Name,
-		"policy_type":        a.Type,
-		"policy_is_private":  a.Private,
-		"policy_selector":    a.Selector,
-	}
 }
